// Front_end/src/app/PartnershipManagement/models/entreprise.ts
// src/app/models/entreprise.ts
export interface Entreprise {
    idEntreprise?: number;
    nameEntreprise: string;
    addressEntreprise: string;
    phoneEntreprise: string;
    emailEntreprise: string;
    descriptionEntreprise: string;
    partner?: any; // Replace with proper User model if available
    partnerships?: any[]; // Replace with proper Partnership model if available
    banned?: boolean;
    scoreEntreprise?: number;
    partnerId?: number;
    partnerUsername?: string;

  }